requirements.txt
aio-pika==9.5.8
aiormq==6.9.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.10.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
colorama==0.4.6
coverage==7.12.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.2
flake8==7.3.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
multidict==6.7.0
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pybreaker==1.0.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
